# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Intania Shop API
  version: 1.0.0
  description: |
    Backend API for the ESC Chula Intania Shop.
    Identity comes from a JWT. Operations described as Admin only require the ADMIN role.
servers:
  - url: /
    description: Same host as this documentation
...
paths:
...
  /auth/google:
    get:
      operationId: startGoogleOAuth
      tags: [Authentication]
      summary: Start Google OAuth with PKCE
      description: Returns the Google authorization URL and sets an OAuth state cookie.
      responses:
        "200": { $ref: "#/components/responses/AuthStart" }
        "500": { $ref: "#/components/responses/Error" }
  /auth/google/redirect:
    get:
      operationId: redirectToGoogleOAuth
      tags: [Authentication]
      summary: Redirect to Google OAuth
      responses:
        "307":
          description: Temporary redirect to Google
          headers:
            Location:
              schema: { type: string, format: uri }
        "500": { $ref: "#/components/responses/Error" }
  /auth/google/callback:
    get:
      operationId: completeGoogleOAuth
      tags: [Authentication]
      summary: Complete Google OAuth and issue a JWT
      parameters:
        - { name: code, in: query, schema: { type: string }, description: Required on a successful callback }
        - { name: state, in: query, schema: { type: string }, description: Required on a successful callback }
        - { name: error, in: query, schema: { type: string } }
        - { name: error_description, in: query, schema: { type: string } }
      responses:
        "200": { $ref: "#/components/responses/Login" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "500": { $ref: "#/components/responses/Error" }
        "502": { $ref: "#/components/responses/Error" }
...
  /cart:
    get:
      operationId: getCart
      tags: [Cart]
      summary: Get the caller cart
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/Cart" }
        "401": { $ref: "#/components/responses/Error" }
        "500": { $ref: "#/components/responses/Error" }
  /cart/items:
    put:
      operationId: addCartItem
      tags: [Cart]
      summary: Add an item to the caller cart
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddCartItemRequest" }
      responses:
        "200": { $ref: "#/components/responses/AddCartItem" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
...
  /promos:
    get:
      operationId: listPromos
      tags: [Content]
      summary: List promotions
      responses:
        "200": { $ref: "#/components/responses/PromoList" }
        "500": { $ref: "#/components/responses/Error" }
    post:
      operationId: createPromo
      tags: [Content]
      summary: Create a promotion
      description: Admin only.
      security: [{ bearerAuth: [] }]
      requestBody: { $ref: "#/components/requestBodies/CreateContent" }
      responses:
        "201": { $ref: "#/components/responses/Promo" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /promos/active:
    get:
      operationId: listActivePromos
      tags: [Content]
      summary: List active promotions
      responses:
        "200": { $ref: "#/components/responses/PromoList" }
        "500": { $ref: "#/components/responses/Error" }
  /promos/{id}:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    get:
      operationId: getPromo
      tags: [Content]
      summary: Get a promotion
      responses:
        "200": { $ref: "#/components/responses/Promo" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    put:
      operationId: updatePromo
      tags: [Content]
      summary: Update a promotion
      description: Admin only.
      security: [{ bearerAuth: [] }]
      requestBody: { $ref: "#/components/requestBodies/UpdateContent" }
      responses:
        "200": { $ref: "#/components/responses/Promo" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deletePromo
      tags: [Content]
      summary: Delete a promotion
      description: Admin only.
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Promotion deleted }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /banners:
    get:
      operationId: listBanners
      tags: [Content]
      summary: List banners
      responses:
        "200": { $ref: "#/components/responses/BannerList" }
        "500": { $ref: "#/components/responses/Error" }
    post:
      operationId: createBanner
      tags: [Content]
      summary: Create a banner
      description: Admin only.
      security: [{ bearerAuth: [] }]
      requestBody: { $ref: "#/components/requestBodies/CreateContent" }
      responses:
        "201": { $ref: "#/components/responses/Banner" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /banners/active:
    get:
      operationId: listActiveBanners
      tags: [Content]
      summary: List active banners
      responses:
        "200": { $ref: "#/components/responses/BannerList" }
        "500": { $ref: "#/components/responses/Error" }
  /banners/{id}:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    get:
      operationId: getBanner
      tags: [Content]
      summary: Get a banner
      responses:
        "200": { $ref: "#/components/responses/Banner" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    put:
      operationId: updateBanner
      tags: [Content]
      summary: Update a banner
      description: Admin only.
      security: [{ bearerAuth: [] }]
      requestBody: { $ref: "#/components/requestBodies/UpdateContent" }
      responses:
        "200": { $ref: "#/components/responses/Banner" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteBanner
      tags: [Content]
      summary: Delete a banner
      description: Admin only.
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Banner deleted }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
...
  requestBodies:
...
    CreateContent:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/CreateContentRequest" }
    UpdateContent:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ContentInput" }
...
  responses:
    Error:
      description: Request failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    AuthStart: { description: OAuth flow started, content: { application/json: { schema: { $ref: "#/components/schemas/AuthStartEnvelope" } } } }
    Login: { description: Authentication completed, content: { application/json: { schema: { $ref: "#/components/schemas/LoginEnvelope" } } } }
...
    Cart: { description: Caller cart, content: { application/json: { schema: { $ref: "#/components/schemas/CartEnvelope" } } } }
    AddCartItem: { description: Cart item added, content: { application/json: { schema: { $ref: "#/components/schemas/AddCartItemEnvelope" } } } }
...
    Promo: { description: Promotion, content: { application/json: { schema: { $ref: "#/components/schemas/PromoEnvelope" } } } }
    PromoList: { description: Promotions, content: { application/json: { schema: { $ref: "#/components/schemas/PromoListEnvelope" } } } }
    Banner: { description: Banner, content: { application/json: { schema: { $ref: "#/components/schemas/BannerEnvelope" } } } }
    BannerList: { description: Banners, content: { application/json: { schema: { $ref: "#/components/schemas/BannerListEnvelope" } } } }
...
  schemas:
    SuccessBase:
      type: object
      required: [success]
      properties:
        success: { type: boolean, const: true }
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, const: false }
        error: { type: string, example: Product not found }
    User:
      type: object
      required: [id, full_name, email, role, profile_picture]
      properties:
        id: { type: integer, format: int64 }
        full_name: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [USER, ADMIN] }
        profile_picture: { type: [string, "null"], format: uri }
...
    AddCartItemRequest:
      type: object
      required: [variant_id, quantity]
      properties:
        variant_id: { type: integer, format: int64, minimum: 1 }
        quantity: { type: integer, format: int32, minimum: 1 }
    CartItem:
      type: object
      properties:
        item_id: { type: integer, format: int64 }
        cart_id: { type: integer, format: int64 }
        variant_id: { type: integer, format: int64 }
        quantity: { type: integer, format: int32 }
    Cart:
      type: object
      required: [cart_id, user_id, items]
      properties:
        cart_id: { type: integer, format: int64 }
        user_id: { type: integer, format: int64 }
        items:
          type: array
          items:
            type: object
            properties:
              item_id: { type: integer, format: int64 }
              cart_id: { type: integer, format: int64 }
              variant_id: { type: integer, format: int64 }
              quantity: { type: integer, format: int32 }
              available_stock: { type: [integer, "null"], format: int32 }
              variant: { type: object, additionalProperties: true }
...
    ContentInput:
      type: object
      properties:
        img_url: { type: string, format: uri }
        link_url: { type: [string, "null"], format: uri }
        is_active: { type: boolean }
    CreateContentRequest:
      allOf:
        - { $ref: "#/components/schemas/ContentInput" }
        - { type: object, required: [img_url] }
    Promo:
      type: object
      properties:
        promo_id: { type: integer, format: int64 }
        img_url: { type: string, format: uri }
        link_url: { type: [string, "null"], format: uri }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
    Banner:
      type: object
      properties:
        banner_id: { type: integer, format: int64 }
        img_url: { type: string, format: uri }
        link_url: { type: [string, "null"], format: uri }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
...
    AuthStartEnvelope:
      allOf:
        - { $ref: "#/components/schemas/SuccessBase" }
        - { type: object, required: [data], properties: { data: { type: object, required: [auth_url], properties: { auth_url: { type: string, format: uri } } } } }
    LoginEnvelope:
      allOf:
        - { $ref: "#/components/schemas/SuccessBase" }
        - { type: object, required: [data], properties: { data: { type: object, required: [user, token, message], properties: { user: { $ref: "#/components/schemas/User" }, token: { type: string }, message: { type: string } } } } }
...
    CartEnvelope:
      allOf:
        - { $ref: "#/components/schemas/SuccessBase" }
        - { type: object, required: [data], properties: { data: { $ref: "#/components/schemas/Cart" } } }
    AddCartItemEnvelope:
      allOf:
        - { $ref: "#/components/schemas/SuccessBase" }
        - { type: object, required: [data], properties: { data: { type: object, properties: { item: { $ref: "#/components/schemas/CartItem" }, message: { type: string } } } } }
...
    PromoEnvelope:
      allOf:
        - { $ref: "#/components/schemas/SuccessBase" }
        - { type: object, required: [data], properties: { data: { $ref: "#/components/schemas/Promo" } } }
    PromoListEnvelope:
      allOf:
        - { $ref: "#/components/schemas/SuccessBase" }
        - { type: object, required: [data], properties: { data: { type: array, items: { $ref: "#/components/schemas/Promo" } } } }
    BannerEnvelope:
      allOf:
        - { $ref: "#/components/schemas/SuccessBase" }
        - { type: object, required: [data], properties: { data: { $ref: "#/components/schemas/Banner" } } }
    BannerListEnvelope:
      allOf:
        - { $ref: "#/components/schemas/SuccessBase" }
        - { type: object, required: [data], properties: { data: { type: array, items: { $ref: "#/components/schemas/Banner" } } } }
...